add sample project

+++ LoadUnManageDLL/Program.cs
@@ -1,9 +1,9 @@
 ﻿using System;
-using System.Collections;
-using System.ComponentModel;
 using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
+using PInvoke;
+using Win32Exception = System.ComponentModel.Win32Exception;
 
 namespace LoadUnManageDLL
 {
@@ -20,24 +20,24 @@ namespace LoadUnManageDLL
         static void UnzipAndLoad()
         {
             var folderPath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            var dllPath = Path.Combine(folderPath, $"{nameof(Resource.pdfium)}.dll");
-            File.WriteAllBytes(dllPath,Resource.pdfium);
-            LoadDll(dllPath);
+            var dllPath = Path.Combine(folderPath, $"{nameof(Resource.pdfium)}.dll");//解压输出的路径
+            if (!File.Exists(dllPath))
+                File.WriteAllBytes(dllPath, Resource.pdfium);
+            LoadDll(dllPath);//应该每次都加载非托管
         }
 
-        ///// <summary>.
-        ///// managed wrapper around LoadLibrary
-        ///// </summary>
-        ///// <param name="dllName"></param>
+        /// <summary>
+        /// 加载非托管DLL
+        /// </summary>
+        /// <param name="dllName"></param>
         public static void LoadDll(string dllName)
         {
-            IntPtr h = LoadLibrary(dllName);
-            if (h == IntPtr.Zero)
+            var h =Kernel32.LoadLibrary(dllName);
+            if (h.IsInvalid)//是否是无效的
             {
                 Exception e = new Win32Exception();
                 throw new DllNotFoundException($"Unable to load library: {dllName}", e);
             }
-
             Console.WriteLine("Load library successful");
         }
